% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cambodia_births_census.R
\docType{data}
\name{cambodia_births_census}
\alias{cambodia_births_census}
\title{Estimates of births in Cambodia from the 2008 Cambodian census}
\format{An array with dimensions "age" and "province".}
\source{
Calculated from data downloaded from the Cambodia General Population
Census database on the Latin American and Caribbean Demographic
Center (CELADE) website, on November 21, 2017.
}
\usage{
cambodia_births_census
}
\description{
Estimated numbers of births, by province and age of mother,
from the 2008 Cambodian population census.
}
\keyword{datasets}
