% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cambodia_births_dhs.R
\docType{data}
\name{cambodia_births_dhs}
\alias{cambodia_births_dhs}
\title{Estimates of births in Cambodia from the 2010 Cambodia Demographic and Health Survey}
\format{A data frame, with three columns:
\describe{
  \item{age}{Age of mother}
  \item{births}{Estimated number of births}
  \item{se}{Standard error of the estimate}
}}
\source{
National Institute of Statistics, Directorate General for Health, and ICF (2011)
Cambodia Demographic and Health Survey 2010 [Dataset] KHIR61FL.DTA,
National Institute of Statistics, Directorate General for Health, and ICF Macro.
}
\usage{
cambodia_births_dhs
}
\description{
The estimates were calculated from individual-level data, using design-based
methods implemented in package \code{survey}.
}
\details{
The individual-level data are confidential, and are not included in package
\code{bdefdata}.
}
\keyword{datasets}
