% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cambodia_births_dhs_revised.R
\docType{data}
\name{cambodia_births_dhs_revised}
\alias{cambodia_births_dhs_revised}
\title{Estimates of births in Cambodia from the 2010 Cambodia Demographic and Health Survey,
with ages 40-44 and 45-49 combined}
\format{An object of class \code{data.frame} with 6 rows and 3 columns.}
\usage{
cambodia_births_dhs_revised
}
\description{
As for \code{\link{cambodia_births_dhs}}, but with the final two age groups
aggregated into a single age group 40-49.
}
\keyword{datasets}
