% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cambodia_population.R
\docType{data}
\name{cambodia_population}
\alias{cambodia_population}
\title{Estimates of population in Cambodia from the 2008 Cambodian census}
\format{An array with dimensions "age", "sex", and "province".}
\source{
Calculated from data downloaded from the Cambodia General Population
Census database on the Latin American and Caribbean Demographic
Center (CELADE) website, on November 21, 2017.
}
\usage{
cambodia_population
}
\description{
Estimated population by age, sex, and province, 
from the 2008 Cambodian population census.
}
\keyword{datasets}
