% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cambodia_poverty.R
\docType{data}
\name{cambodia_poverty}
\alias{cambodia_poverty}
\title{Estimates of provincial poverty rates Cambodia, 2009}
\format{A data frame, with two columns:
\describe{
  \item{province}{Province of Cambodia}
  \item{poverty}{Percent below poverty line}
}}
\source{
Asian Development Bank. 2014. Cambodia Country Poverty Analysis 2014.
Manila: Asian Development Bank
}
\usage{
cambodia_poverty
}
\description{
Estimates of the percent of the population living below the poverty line,
by province, compiled by the Asian Development Bank.
}
\keyword{datasets}
