% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/china_deaths_un.R
\docType{data}
\name{china_deaths_un}
\alias{china_deaths_un}
\title{Deaths (hundreds) in China, as estimated by the United Nations}
\format{An array with dimensions "age" and "time".}
\source{
Table "Number of deaths by age and sex (thousands)"
in the United Nations Population Division online database
"World Population Prospects".
Data  downloaded on 2 December 2017.
}
\usage{
china_deaths_un
}
\description{
Estimates of deaths for the period 1990-2015, constructed
by UN demographers. The counts are shown in multiples of 100.
}
\keyword{datasets}
