% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/china_population_census.R
\docType{data}
\name{china_population_census}
\alias{china_population_census}
\title{Population counts (hundreds) from Chinese censuses, by age}
\format{An array with dimensions "age" and "time".}
\source{
Table "Population by age, sex and urban/rural residence"
in the United Nations Statistics Division online database
"Population Censuses Datasets (1995 - Present)".
Data  downloaded on 21 January 2018.
}
\usage{
china_population_census
}
\description{
Counts of respondents from the 1990, 2000, and 2010 censuses.
The counts have been adjusted by UN demographers. The counts
are shown in multiples of 100.
}
\keyword{datasets}
