% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iceland_migration.R
\docType{data}
\name{iceland_migration}
\alias{iceland_migration}
\title{Internal migration in Iceland}
\format{An array with dimensions "age", "sex",
"region_orig", "region_dest", and "time".}
\source{
Table "Internal migration between regions
by sex and age 1986-2016 - Division into municipalities as
of 1 January 2017" on the Statistics Iceland website
(downloaded on 17 December 2017).
}
\usage{
iceland_migration
}
\description{
The number of moves between regions of Iceland,
by age, sex, region of origin, region of destination,
and time.
}
\keyword{datasets}
