% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iceland_population.R
\docType{data}
\name{iceland_population}
\alias{iceland_population}
\title{Population of Iceland}
\format{An array with dimensions "age", "sex", "region", and "time".}
\source{
Table "Population by municipality, age and sex 1998-2017
- Division into municipalities as of 1 January 2017"
on the Statistics Iceland website
(downloaded on 17 December 2017).
}
\usage{
iceland_population
}
\description{
The population of Icleand, by age, sex, region, and time.
}
\keyword{datasets}
