% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maori_deaths.R
\docType{data}
\name{maori_deaths}
\alias{maori_deaths}
\title{Deaths of Maori people in New Zealand}
\format{An array with dimensions "age", "sex", and "year".}
\source{
Table "Deaths by age and sex (Maori population) (Annual-Dec)"
in the online database "Infoshare", on the Statistics New Zealand
website. Downloaded on 30 June 2018.
}
\usage{
maori_deaths
}
\description{
An array giving counts of deaths of Maori people,
for calendar years 1996-2017. Maori are the indigenous people
of New Zealand.
}
\details{
The counts have been randomly rounded to base 3, to protect
confidentiality.

For an analysis of these data, see Chapter 3 of
\emph{Bayesian Demographic Estimation and Forecasting}.
For a discussion of random rounding to base 3, see Chapter
15.
}
\keyword{datasets}
