% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maori_popn.R
\docType{data}
\name{maori_popn}
\alias{maori_popn}
\title{Population of Maori people in New Zealand}
\format{An array with dimensions "age", "sex", and "year".}
\source{
Table "Estimated resident population (ERP),
national population by ethnic group, age, and sex,
30 June 1996, 2001, 2006, and 2013"
in the online database "Infoshare", on the Statistics New Zealand
website. Downloaded on 2 July 2018.
}
\usage{
maori_popn
}
\description{
An array giving population counts for Maori people,
at 30 June 1996, 30 June 2001, 30 June 2006, and 30 June 2013.
Maori are the indigenous people of New Zealand.
}
\details{
The counts have been rounded to base 10.

For an analysis of these data, see Chapter 3 of
\emph{Bayesian Demographic Estimation and Forecasting}.
}
\keyword{datasets}
