% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netherlands_exposure.R
\docType{data}
\name{netherlands_exposure}
\alias{netherlands_exposure}
\title{Estimated and projected person-years lived in the Netherlands,
by age and sex, 2003-2021}
\format{An array with dimensions "age", "sex", and "time".}
\source{
Calculated from tables "Population on 1 January (tps00001)"
and "Population projections (tps00002)" on the Eurostat database.
Data downloaded on 26 November 2017.
}
\usage{
netherlands_exposure
}
\description{
Estimated and projected person-years lived in the Netherlands,
by age and sex, 2003-2021
}
\keyword{datasets}
