% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netherlands_population.R
\docType{data}
\name{netherlands_population}
\alias{netherlands_population}
\title{Estimated and projected population of the Netherlands,
by age and sex, 2003-2021}
\format{An array with dimensions "age", "sex", and "time".}
\source{
Table "Historical population data and projections
(1950-2050)" from OECD database 'OECD.Stat'.
Data downloaded 5 June 2016.
}
\usage{
netherlands_population
}
\description{
Estimated and projected population of the Netherlands,
by age and sex, 2003-2021
}
\keyword{datasets}
