% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_zealand_address_admin.R
\docType{data}
\name{new_zealand_address_admin}
\alias{new_zealand_address_admin}
\title{Address changes recorded in administrative data, New Zealand, 2007-2016}
\format{An array with dimensions "region_orig",
"region_dest", and "time".}
\source{
Custom tabulation from Stats NZ.
}
\usage{
new_zealand_address_admin
}
\description{
Address changes, by region of origin and region of destination,
for 16 New Zealand regions, 2007-2016.
}
\details{
The data come from linked individual-level datasets in the
'Integrated Data Infrastructure' at Stats NZ, the New Zealand
statistics agency.  The Stats NZ staff preparing the tabulations accessed
data in accordance with the security and confidentiality provisions
of the Statistics Act 1975. Only people authorized by the Statistics
Act 1975 are allowed to see data about a particular person,
household, business, or organization.
}
\keyword{datasets}
