% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_zealand_arrivals.R
\docType{data}
\name{new_zealand_arrivals}
\alias{new_zealand_arrivals}
\title{Permament and long-term arrivals to New Zealand,
1997-2016.}
\format{An array with dimensions "region" and "time"}
\source{
Custom tabulation from Stats NZ.
}
\usage{
new_zealand_arrivals
}
\description{
A permanent and long-term arrival is border crossing by a
person who intends to reside in the country for at least on year.
}
\details{
The years are June years.  For instance, the year "1997" runs from 1 July 1996
to 30 Jun 1997.

The data derive from international arrival cards filled
out when entering the country.
}
\keyword{datasets}
