% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_zealand_births_reg.R
\docType{data}
\name{new_zealand_births_reg}
\alias{new_zealand_births_reg}
\title{Registered births in New Zealand by region, 1997-2016.}
\format{An array with dimensions "region" and "year"}
\source{
Custom tabulation from Statistics New Zealand.
}
\usage{
new_zealand_births_reg
}
\description{
The years are June years.  For instance, the year "1997" runs from 1 July 1996
to 30 Jun 1997.
}
\details{
To preserve confidentiality, birth counts have been randomly rounded to base 3.
}
\keyword{datasets}
