% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_zealand_departures.R
\docType{data}
\name{new_zealand_departures}
\alias{new_zealand_departures}
\title{Permament and long-term departures from New Zealand,
1997-2016.}
\format{An array with dimensions "region" and "time"}
\source{
Custom tabulation from Stats NZ.
}
\usage{
new_zealand_departures
}
\description{
A permanent and long-term departure is a border crossing
by a person who intends to reside out of the country for at least
one year.
}
\details{
The years are June years.  For instance, the year "1997" runs from 1 July 1996
to 30 Jun 1997.

The data derive from international departure cards filled
out when leaving the country.
}
\keyword{datasets}
