% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_zealand_internal_census.R
\docType{data}
\name{new_zealand_internal_census}
\alias{new_zealand_internal_census}
\title{Address changes in census data, New Zealand, 2008-2013}
\format{An array with dimensions "region_orig",
"region_dest", and "time".}
\source{
Custom tabulation from Stats NZ.
}
\usage{
new_zealand_internal_census
}
\description{
Counts of respondents from the 2013 census by their region of
residence at the time of the census (denoted "region_dest"),
and their regeion of residence in 2008 (denoted "region_orig").
}
\details{
The counts were obtained from a dataset in which values had been
imputed and rounded to base 3, and differ slightly from
published tabulations.
}
\keyword{datasets}
