% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_zealand_popn_admin.R
\docType{data}
\name{new_zealand_popn_admin}
\alias{new_zealand_popn_admin}
\title{Estimates of New Zealand population from administrative
data, 2007-2016.}
\format{An array with dimension "time"}
\source{
Statistics New Zealand website, downloaded 7 December 2017.
}
\usage{
new_zealand_popn_admin
}
\description{
The population counts refer to 30 June of each year. The data are
"Experimental population estimates from linked administrative data",
published in 2 October 2017.
}
\keyword{datasets}
