% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_zealand_popn_census.R
\docType{data}
\name{new_zealand_popn_census}
\alias{new_zealand_popn_census}
\title{Estimates of New Zealand population by region,
in census years 1996, 2001, 2006, and 2013}
\format{An array with dimensions "region", and "time"}
\source{
Custom tabulation from Statistics New Zealand.
}
\usage{
new_zealand_popn_census
}
\description{
The population estimates refer to 30 June of each year.
The estimates constructed by starting from census counts,
adjusting for under-coverage and
over-coverage, and then adjusting for births, deaths, and migration
between census day and 30 June.  New Zealand held a census in 2013
rather than 2011 because of an earthquake in 2011.
}
\keyword{datasets}
