% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portugal_exposure.R
\docType{data}
\name{portugal_exposure}
\alias{portugal_exposure}
\title{Years of exposure in Portugal by age, sex, and calendar year.}
\format{An array with dimensions "age", "sex" and "year".}
\source{
Human Mortality Database. University of California, Berkeley (USA),
and Max Planck Institute for Demographic Research (Germany).
Available at \url{www.mortality.org} or \url{www.humanmortality.de}
(data downloaded on 8 October 2017).
}
\usage{
portugal_exposure
}
\description{
The data are from the Human Mortality Database.
More recent versions, and a comprehensive description of the data,
are available at the HMD website.
}
\keyword{datasets}
