% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweden_deaths.R
\docType{data}
\name{sweden_deaths}
\alias{sweden_deaths}
\title{Counts of Infant Deaths in Sweden by County and Year, 1995-2015}
\format{An array with dimensions "age", "sex", "region", and "year".}
\source{
Database "Deaths by region, age (during the year) and sex.
Year 1968 - 2017". Downloaded from Statistics Sweden on 13 July 2018.
}
\usage{
sweden_deaths
}
\description{
The data are from the Statistics Sweden website.
}
\details{
Accents have been and umlauts have been removed from county names
to avoid problems with encoding. (Apologies to Swedish speakers.)
}
\keyword{datasets}
