% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twelve_emigration.R
\docType{data}
\name{twelve_emigration}
\alias{twelve_emigration}
\title{Emigration in the population of twelve fictitious individuals}
\format{A data.frame with columns "name", "time", "age", and "region".}
\usage{
twelve_emigration
}
\description{
A data.frame giving information on emigration in the population
of 12 fictitious individuals.
}
\details{
For more on the twelve fictitious individuals, see Chapter 3 of
\emph{Bayesian Demographic Estimation and Forecasting}.
}
\keyword{datasets}
