% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twelve_movements.R
\docType{data}
\name{twelve_movements}
\alias{twelve_movements}
\title{Migrations in the population of twelve fictitious individuals}
\format{A data.frame with columns "name", "time", "age", "region_orig",
and "region_dest".}
\usage{
twelve_movements
}
\description{
A data.frame giving information on internal migration by the 12 fictitious
individuals.
}
\details{
For more on the twelve fictitious individuals, see Chapter 3 of
\emph{Bayesian Demographic Estimation and Forecasting}.
}
\keyword{datasets}
