% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twelve_population.R
\docType{data}
\name{twelve_population}
\alias{twelve_population}
\title{Individuals from the population of twelve fictitious individuals
who are alive at the start and end of the observation period}
\format{A data.frame with columns "name", "time", "age", and "region".}
\usage{
twelve_population
}
\description{
A data.frame listing the individuals who are alive the start
and end of the observation of the period,
together with their ages, and regions of residence.
}
\details{
For more on the twelve fictitious individuals, see Chapter 3 of
\emph{Bayesian Demographic Estimation and Forecasting}.
}
\keyword{datasets}
