% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twelve_taxes.R
\docType{data}
\name{twelve_taxes}
\alias{twelve_taxes}
\title{Taxes in the population of twelve fictitious individuals}
\format{A data.frame with columns "name", "time", "age", "region",
and amount.}
\usage{
twelve_taxes
}
\description{
A data.frame giving information on taxes in the population
of 12 fictitious individuals.
}
\details{
For more on the twelve fictitious individuals, see Chapter 3 of
\emph{Bayesian Demographic Estimation and Forecasting}.
}
\keyword{datasets}
