% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netherlands_cpi.R
\docType{data}
\name{netherlands_cpi}
\alias{netherlands_cpi}
\title{Consumer price index in the Netherlands, 1996-2015}
\format{An array with dimension 'time'.}
\source{
Table "Key Short-Term Economic Indicators: Consumer
Prices - Annual inflation"from OECD database 'OECD.Stat'.
Data downloaded 3 June 2016.
}
\usage{
netherlands_cpi
}
\description{
Consumer price index in the Netherlands, 1996-2015
}
\keyword{datasets}
