% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netherlands_exposure.R
\docType{data}
\name{netherlands_exposure}
\alias{netherlands_exposure}
\title{Estimated and projected person-years lived in the Netherlands,
by age and sex, 2003-2021}
\format{An array with dimensions "age", "sex", and "time".}
\source{
Calculated from table
"Historical population data and projections
(1950-2050)" from OECD database 'OECD.Stat'.
Data downloaded 5 June 2016.
}
\usage{
netherlands_exposure
}
\description{
Estimated and projected person-years lived in the Netherlands,
by age and sex, 2003-2021
}
\keyword{datasets}
