% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netherlands_health.R
\docType{data}
\name{netherlands_health}
\alias{netherlands_health}
\title{Health expenditure in the Netherlands, 2003-2011}
\format{An array with dimensions  "age", "sex", and "time".}
\source{
Table "Current Health Spending by Age" in the OECD database
'OECD.stat'. Data downloaded on 24 November 2017.
}
\usage{
netherlands_health
}
\description{
Total health expenditures in millions of Euros, for all providers,
by age, sex, and year.
}
\details{
Data are only available for the years 2003, 2005, 2007, and 2011.
Missing values are used for intermediate years.
}
\keyword{datasets}
